jsp page now uses the solar output class

+++ SolarPowerCalculator_AppEngine/src/environmentalSpecifications/LocationDetails.java
@@ -10,15 +10,15 @@ import solarPowerCalculator.CalculatorException;
 public class LocationDetails {
 
 	// Class Variables
-	private int daylightHours; // Hours
-	private static int hoursInDay = 24; //Hours
+	private double daylightHours; // Hours
+	private static double hoursInDay = 24; //Hours
 	
 	/**
 	 * Initializes a new instance of location details with specified daylight exposure
 	 * @param daylightHours hours of daylight system is exposed to
 	 * @throws CalculatorException if daylightHours exceeds hoursInDay
 	 */
-	public LocationDetails (int daylightHours) throws CalculatorException{
+	public LocationDetails (double daylightHours) throws CalculatorException{
 		
 		if (daylightHours > hoursInDay || daylightHours < 0)
 			throw new CalculatorException("daylightHours exceeds limits");
@@ -31,7 +31,7 @@ public class LocationDetails {
 	 *  Gets the number of hours of daylight in the day
 	 * @return hours of daylight in a day
 	 */
-	public int getDaylightHours () {
+	public double getDaylightHours () {
 		return daylightHours;
 	}
